[patch] refactor employees context
