+++ lib/ehealth/web/controllers/employee_request_controller.ex
@@ -3,15 +3,15 @@ defmodule EHealth.Web.EmployeeRequestController do
 
   use EHealth.Web, :controller
   alias Scrivener.Page
-  alias EHealth.Employee.API
+  alias EHealth.EmployeeRequests, as: API
   alias EHealth.API.Mithril
-  alias EHealth.Employee.Request
+  alias EHealth.EmployeeRequests.EmployeeRequest, as: Request
   alias EHealth.PRM.LegalEntities
 
   action_fallback EHealth.Web.FallbackController
 
   def index(conn, params) do
-    with {%Page{} = paging, references} <- API.list_employee_requests(params) do
+    with {%Page{} = paging, references} <- API.list(params) do
       render(conn, "index.json",
         employee_requests: paging.entries,
         references: references,
@@ -23,7 +23,7 @@ defmodule EHealth.Web.EmployeeRequestController do
   def create(%Plug.Conn{req_headers: req_headers} = conn, params) do
     client_id = get_client_id(req_headers)
     params = put_in(params, ["employee_request", "legal_entity_id"], client_id)
-    with {:ok, employee_request} <- API.create_employee_request(params) do
+    with {:ok, employee_request} <- API.create(params) do
       render(conn, "show.json",
         employee_request: employee_request,
         legal_entity: get_legal_entity(employee_request)
@@ -32,7 +32,7 @@ defmodule EHealth.Web.EmployeeRequestController do
   end
 
   def show(conn, %{"id" => id}) do
-    employee_request = API.get_employee_request_by_id!(id)
+    employee_request = API.get_by_id!(id)
 
     conn
     |> put_urgent_user_id(employee_request)
@@ -50,7 +50,7 @@ defmodule EHealth.Web.EmployeeRequestController do
 
   def approve(%Plug.Conn{req_headers: req_headers} = conn, %{"id" => id}) do
     with :ok <- API.check_employee_request(req_headers, id) do
-      with {:ok, employee_request} <- API.approve_employee_request(id, req_headers) do
+      with {:ok, employee_request} <- API.approve(id, req_headers) do
         render(conn, "show.json",
           employee_request: employee_request,
           legal_entity: get_legal_entity(employee_request)
@@ -61,7 +61,7 @@ defmodule EHealth.Web.EmployeeRequestController do
 
   def reject(%Plug.Conn{req_headers: req_headers} = conn, %{"id" => id}) do
     with :ok <- API.check_employee_request(req_headers, id) do
-      with {:ok, employee_request} <- API.reject_employee_request(id) do
+      with {:ok, employee_request} <- API.reject(id) do
         render(conn, "show.json",
           employee_request: employee_request,
           legal_entity: get_legal_entity(employee_request)
