+++ lib/ehealth/web/controllers/employees_controller.ex
@@ -4,19 +4,19 @@ defmodule EHealth.Web.EmployeeController do
   use EHealth.Web, :controller
 
   alias Scrivener.Page
-  alias EHealth.Employee.API
-  alias EHealth.Employee.EmployeeUpdater
+  alias EHealth.Employees, as: API
+  alias EHealth.Employees.EmployeeUpdater
 
   action_fallback EHealth.Web.FallbackController
 
   def index(conn, params) do
-    with %Page{} = paging <- API.get_employees(params) do
+    with %Page{} = paging <- API.list(Map.put(params, "is_active", true)) do
       render(conn, "index.json", employees: paging.entries, paging: paging)
     end
   end
 
   def show(%Plug.Conn{req_headers: req_headers} = conn, %{"id" => id}) do
-    with {:ok, employee} <- API.get_employee_by_id(id, req_headers) do
+    with {:ok, employee} <- API.get_by_id(id, req_headers) do
       render(conn, "employee.json", employee: employee)
     end
   end
