+++ lib/ehealth/web/controllers/legal_entity_controller.ex
@@ -61,7 +61,7 @@ defmodule EHealth.Web.LegalEntityController do
     end
   end
 
-  defp assign_employee_request_id(conn, %EHealth.Employee.Request{id: id}) do
+  defp assign_employee_request_id(conn, %EHealth.EmployeeRequests.EmployeeRequest{id: id}) do
     assign_urgent(conn, "employee_request_id", id)
   end
   defp assign_employee_request_id(conn, _employee_request_id), do: conn
