+++ lib/ehealth/web/views/employee_view.ex
@@ -7,7 +7,7 @@ defmodule EHealth.Web.EmployeeView do
   alias EHealth.Parties.Party
   alias EHealth.Divisions.Division
   alias EHealth.PRM.LegalEntities.Schema, as: LegalEntity
-  alias EHealth.PRM.Employees.Schema, as: Employee
+  alias EHealth.Employees.Employee
 
   @doctor Employee.type(:doctor)
   @pharmacist Employee.type(:pharmacist)
