+++ test/web/controllers/employees_controller_test.exs
@@ -3,7 +3,7 @@ defmodule EHealth.Web.EmployeesControllerTest do
 
   use EHealth.Web.ConnCase
   alias EHealth.MockServer
-  alias EHealth.PRM.Employees.Schema, as: Employee
+  alias EHealth.Employees.Employee
   alias EHealth.Parties.Party
   alias Ecto.UUID
   alias EHealth.PRMRepo
